# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "go-admin-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "paths": {
        "/dept/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "创建部门",
                "parameters": [
                    {
                        "description": "创建部门",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateDeptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dept/delete/{id}": {
            "delete": {
                "tags": [
                    "部门管理"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dept/info/{id}": {
            "get": {
                "tags": [
                    "部门管理"
                ],
                "summary": "部门详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SysDept"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dept/page": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "部门列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "部门名称",
                        "name": "deptName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "启用状态 0 启用 1 禁用",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/models.SysDept"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dept/update": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "更新部门",
                "parameters": [
                    {
                        "description": "更新部门",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateDeptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictData/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据字典数据"
                ],
                "summary": "创建字典数据",
                "parameters": [
                    {
                        "description": "创建字典数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateDictDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictData/delete/{id}": {
            "delete": {
                "tags": [
                    "数据字典数据"
                ],
                "summary": "删除字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictData/info/{id}": {
            "get": {
                "tags": [
                    "数据字典数据"
                ],
                "summary": "字典数据详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictData/page": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据字典数据"
                ],
                "summary": "字典数据列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/models.SysDictData"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dictData/update": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据字典数据"
                ],
                "summary": "更新字典数据",
                "parameters": [
                    {
                        "description": "更新字典数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateDictDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.SysDept": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deptName": {
                    "description": "部门名称",
                    "type": "string"
                },
                "deptSort": {
                    "description": "显示排序",
                    "type": "integer"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "leader": {
                    "description": "负责人",
                    "type": "string"
                },
                "parentId": {
                    "description": "父部门ID",
                    "type": "integer"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "models.SysDictData": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建者",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dictLabel": {
                    "description": "字典标签",
                    "type": "string"
                },
                "dictSort": {
                    "description": "字典排序",
                    "type": "integer"
                },
                "dictValue": {
                    "description": "字典键值",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "boolean"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "request.CreateDeptRequest": {
            "type": "object",
            "properties": {
                "deptName": {
                    "description": "部门名称",
                    "type": "string"
                },
                "deptSort": {
                    "description": "显示排序",
                    "type": "integer"
                },
                "leader": {
                    "description": "负责人",
                    "type": "string"
                },
                "parentId": {
                    "description": "父部门ID",
                    "type": "integer"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "integer"
                }
            }
        },
        "request.CreateDictDataRequest": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建者",
                    "type": "string"
                },
                "dictLabel": {
                    "description": "字典标签",
                    "type": "string"
                },
                "dictSort": {
                    "description": "字典排序",
                    "type": "integer"
                },
                "dictValue": {
                    "description": "字典键值",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "boolean"
                }
            }
        },
...
        "request.UpdateDeptRequest": {
            "type": "object",
            "properties": {
                "deptName": {
                    "description": "部门名称",
                    "type": "string"
                },
                "deptSort": {
                    "description": "显示排序",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "leader": {
                    "description": "负责人",
                    "type": "string"
                },
                "parentId": {
                    "description": "父部门ID",
                    "type": "integer"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "integer"
                }
            }
        },
        "request.UpdateDictDataRequest": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建者",
                    "type": "string"
                },
                "dictLabel": {
                    "description": "字典标签",
                    "type": "string"
                },
                "dictSort": {
                    "description": "字典排序",
                    "type": "integer"
                },
                "dictValue": {
                    "description": "字典键值",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "boolean"
                }
            }
        },
...
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "list": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "result.HttpResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}